Move orchest-api types in types module
