+++ services/orchest-api/app/app/core/sessions/__init__.py
@@ -12,9 +12,6 @@ the retrieval of such information should be quite straightforward.
 
 """
 from app.core.sessions._core import (
-    InteractiveSessionConfig,
-    NonInteractiveSessionConfig,
-    SessionType,
     cleanup_resources,
     has_busy_kernels,
     launch,
